Swapped controller to resource controller
